refactor(kernel): isolate projection diagnostics
